Update all base Docker image versions
This doesn't really matter too much, but it's a good idea to
occasionally keep these up to date.

I decided to go with a minor lock instead of patch to avoid Docker
removing older patch releases and having that break everything.

In this case having a patch lock isn't important since this image only
controls the environment where the project gets generated under.